push forgotten fix/changes
Is this a revive attempt for TabAUR? maybe.
It's been almost year that we didn't update TabAUR at all (paru reference? /s)

We have been working on different projects since last commits, evolving our coding style and skills,
thus we didn't have time to push any updates or make releases.
I don't know if we gonna come back to TabAUR like we used to in 2023, but we're going to publish the v0.7.0 release
with a feature that yet other AUR helpers still doesn't have it: using the https://github.com/archlinux/aur mirror because of the DDoS attacks :)

Also burnt might be doing a refactor of the codebase, making it more faster/stable.